Foghorn.Dsl.Algebra.Exorcism

Description

Type erasure instances implementing the Exorcise type class for econometric variable types.

The Exorcise type class (defined in Foghorn.Dsl.Algebra.Var.Internal and exported by Foghorn.Dsl.Algebra.Var) provides a systematic way to convert typed structures to type-erased equivalents while preserving type information through singleton evidence where applicable. This module provides instances for all composite types that contain variables, enabling type erasure to work recursively through complex structures.

Type Erasure Pattern

Expand

Most instances follow the functor pattern, using fmap to recursively apply exorcise to nested components:

instance Exorcise (Foo (EVar ...)) (Foo (XVar ...)) where
    exorcise = fmap exorcise

This allows type erasure to propagate automatically through aggregations, operations, subsamples, and other composite structures.

Orphan instances

Exorcise (EVar xsource i d) (XVar xsource) => Exorcise (AggregateComp (EVar xsource i d)) (AggregateComp (XVar xsource))

Instance details

Methods

exorcise :: AggregateComp (EVar xsource i d) -> AggregateComp (XVar xsource)

Exorcise (EVar xsource i d) (XVar xsource) => Exorcise (AlgebraExp (EVar xsource i d)) (AlgebraExp (XVar xsource))

Instance details

Methods

exorcise :: AlgebraExp (EVar xsource i d) -> AlgebraExp (XVar xsource)

Exorcise (DbFunction xvar) (DbFunction xvar)

Instance details

Methods

exorcise :: DbFunction xvar -> DbFunction xvar

Exorcise (EVar xsource i d) (XVar xsource) => Exorcise (Subsample (EVar xsource i d)) (Subsample (XVar xsource))

Instance details

Methods

exorcise :: Subsample (EVar xsource i d) -> Subsample (XVar xsource)

Exorcise (XVar xsource) (XVar xsource)

Instance details

Methods

exorcise :: XVar xsource -> XVar xsource

Exorcise (By xsource i) (XBy (XVar xsource))

Instance details

Methods

exorcise :: By xsource i -> XBy (XVar xsource)

Exorcise (Having xsource i) (XHaving (XVar xsource))

Instance details

Methods

exorcise :: Having xsource i -> XHaving (XVar xsource)

Exorcise (JoinWhere xsource i) (XJoinWhere (XVar xsource))

Instance details

Methods

exorcise :: JoinWhere xsource i -> XJoinWhere (XVar xsource)

Exorcise (Peer xsource i) (XPeer (XVar xsource))

Instance details

Methods

exorcise :: Peer xsource i -> XPeer (XVar xsource)