Foghorn.Dsl.Internal.TH

Description

Internal Template Haskell utilities for constructing AST nodes from strings. These utilities reduce boilerplate when writing Template Haskell metaprograms that generate variable accessors and quasi-quoters.

Documentation

conE_ :: String -> Exp

Construct a constructor expression from a string name.

conE_ Just produces ConE (mkName Just)

conT_ :: String -> Type

Construct a constructor type from a string name.

conT_ Maybe produces ConT (mkName Maybe)

varE_ :: String -> Exp

Construct a variable expression from a string name.

varE_ "x" produces VarE (mkName "x")

varT_ :: String -> Type

Construct a type variable from a string name.

varT_ "a" produces VarT (mkName "a")

varP_ :: String -> Pat

Construct a variable pattern from a string name.

varP_ "x" produces VarP (mkName "x")

funD_ :: String -> [Clause] -> Dec

Construct a function declaration from a string name and clauses.

funD_ "myFunc" [clause] produces FunD (mkName "myFunc") [clause]

litE_ :: String -> Exp

Construct a string literal expression from a string.

litE_ "hello" produces LitE (StringL "hello")

charToString :: Char -> String

Convert a single character to a string (singleton list).

textLitE :: Name -> Exp

Create a typed Text literal expression from a Name.

Extracts the base name of the Name, creates a string literal from it, and adds a type signature indicating it's Text.

Used when generating Text constants in spliced code.