Foghorn.Dsl.Interpreter.ModelNames

Description

Provides functionality for extracting model names from a research project. This is useful for coordinating estimation and output phases in statistical software. This module is generally not needed by end users.

Usage

myProject :: Project XCoficatVar (Panel DealNumber_ AnnouncementDate_) ()
myProject = do
    estimate_ regression1
    estimate_ regression2

modelNames :: [Text]
modelNames = modelNames myProject
-- Returns: ["regression1", "regression2"]

Documentation

modelNames :: forall {k} xsource (i :: k) (es :: [(Type -> Type) -> Type -> Type]) a. Eff ((Stmt xsource i :: (Type -> Type) -> Type -> Type) ': es) a -> Eff es [Text]

Extract all model names from a research project.

Interprets a project program and collects the names of all estimators that appear in estimate_ statements. The order of names matches the order of estimation in the project.

This is used by tableDisplay_ and tableSave_ to generate the list of model names for the table output.