Foghorn.Dsl.Interpreter.VariablesUsed

Description

Provides functionality for extracting all variables used in a research project. This ensures that generating data management processes are complete. This module is generally not needed by end users.

Usage

myProject :: Project XCoficatVar (Panel DealNumber_ AnnouncementDate_) ()
myProject = do
    [observeQ| debugVar |]
    estimate_ myRegression

allVars :: Set (SpecVar XCoficatVar (Panel DealNumber_ AnnouncementDate_))
allVars = variablesUsed myProject
-- Contains: debugVar + all variables from myRegression

Project variables

variablesUsed :: forall {k1} {k2} xsource (i :: k1) (r :: k2) a. Ord xsource => Eff '[Stmt xsource i :: (Type -> Type) -> Type -> Type] a -> Set (SpecVar xsource i)

Extract all variables used in a research project.

Interprets a project program and collects all variables that appear in:

Estimators (via estimatorVariablesUsed)

Observe statements

Other statement types (comments, headers, etc.) do not contribute variables.