Foghorn.Dsl.Language.Statement

Description

Defines the Stmt Effectful effect GADT that represents atomic statements in a research project. This is the low-level representation used internally by the project DSL.

Most users will not interact with this module directly. Instead, use the high-level functions from Foghorn.Dsl.Language like estimate_, observe_, comment_, etc.

The makeEffect Template Haskell call generates smart constructors for each statement type (suffixed with __).

Documentation

data Stmt xsource (i :: k) (m :: k1) a where

Research project statement effect.

Represents the atomic operations that can appear in a research project. This is an Effectful effect GADT where each constructor corresponds to a different kind of statement.

The type parameters are:

xsource - Variable source (e.g., XCoficatVar)

i - Index structure (e.g., 'Panel DealNumber_ AnnouncementDate_')

m - Monadic context (managed by Effectful)

a - Return type of the statement

Constructors

Expand

BlankLine__ - Insert a blank line

Comment__ - Add a comment

Divider__ - Insert a divider line

Estimate__ - Run an estimator

Table__ - Output estimation results table

Header__ - Add a section header

Observe__ - Display variable values for debugging

ReadCsv__ - Load CSV data

Raw__ - Insert raw code

Constructors

BlankLine__ :: forall {k} {k1} xsource (i :: k) (m :: k1). Stmt xsource i m ()

Comment__ :: forall {k} {k1} xsource (i :: k) (m :: k1). Text -> Stmt xsource i m ()

Divider__ :: forall {k} {k1} xsource (i :: k) (m :: k1). Stmt xsource i m ()

Estimate__ :: forall {k} {k1} xsource (i :: k) a1 (m :: k1). Estimator xsource i a1 => a1 -> Stmt xsource i m ()

Table__ :: forall {k} {k1} xsource (i :: k) (m :: k1). Maybe FilePath -> [Text] -> Stmt xsource i m ()

Header__ :: forall {k} {k1} xsource (i :: k) (m :: k1). Text -> Stmt xsource i m ()

Observe__ :: forall {k} {k1} xsource (i :: k) (m :: k1). SpecVar xsource i -> Stmt xsource i m ()

ReadCsv__ :: forall {k} {k1} xsource (i :: k) (m :: k1). Text -> Stmt xsource i m ()

Raw__ :: forall {k} {k1} xsource (i :: k) (m :: k1). Text -> Stmt xsource i m ()

raw__ :: forall {k1} {k2} xsource (i :: k1) (es :: [Effect]). (HasCallStack, (Stmt xsource i :: (Type -> Type) -> Type -> Type) :> es) => Text -> Eff es ()

Perform the operation Raw__.

readCsv__ :: forall {k1} {k2} xsource (i :: k1) (es :: [Effect]). (HasCallStack, (Stmt xsource i :: (Type -> Type) -> Type -> Type) :> es) => Text -> Eff es ()

Perform the operation ReadCsv__.

observe__ :: forall {k1} {k2} xsource (i :: k1) (es :: [Effect]). (HasCallStack, (Stmt xsource i :: (Type -> Type) -> Type -> Type) :> es) => SpecVar xsource i -> Eff es ()

Perform the operation Observe__.

header__ :: forall {k1} {k2} xsource (i :: k1) (es :: [Effect]). (HasCallStack, (Stmt xsource i :: (Type -> Type) -> Type -> Type) :> es) => Text -> Eff es ()

Perform the operation Header__.

table__ :: forall {k1} {k2} xsource (i :: k1) (es :: [Effect]). (HasCallStack, (Stmt xsource i :: (Type -> Type) -> Type -> Type) :> es) => Maybe FilePath -> [Text] -> Eff es ()

Perform the operation Table__.

estimate__ :: forall {k1} {k2} xsource (i :: k1) a (es :: [Effect]). (HasCallStack, (Stmt xsource i :: (Type -> Type) -> Type -> Type) :> es, Estimator xsource i a) => a -> Eff es ()

Perform the operation Estimate__.

divider__ :: forall {k1} {k2} xsource (i :: k1) (es :: [Effect]). (HasCallStack, (Stmt xsource i :: (Type -> Type) -> Type -> Type) :> es) => Eff es ()

Perform the operation Divider__.

comment__ :: forall {k1} {k2} xsource (i :: k1) (es :: [Effect]). (HasCallStack, (Stmt xsource i :: (Type -> Type) -> Type -> Type) :> es) => Text -> Eff es ()

Perform the operation Comment__.

blankLine__ :: forall {k1} {k2} xsource (i :: k1) (es :: [Effect]). (HasCallStack, (Stmt xsource i :: (Type -> Type) -> Type -> Type) :> es) => Eff es ()

Perform the operation BlankLine__.