Foghorn.Dsl.Language.Types.Spec.Lenses

Description

Provides lens type classes and convenience functions for manipulating estimator specifications in a generic way. These lenses allow you to modify components (LHS, RHS, conditions, etc.) of any estimator without knowing the concrete estimator type.

The module includes:

Lens type classes - HasName, HasLhs, HasRhs, HasConditions, HasVce, HasAbsorb

Setter functions - Convenient ways to set values (e.g., name_, conditions_)

Transformation functions - Generic variable transformations (e.g., mapMaybeLhsVar)

Most users will not import this module directly. Instead, import Foghorn.Dsl.Language.Types which re-exports these utilities along with core data structures and the estimator abstraction.

Name lens

class HasName spec where

Lens type class for estimator names.

Methods

nameL :: Lens' spec Text

name_ :: HasName spec => Text -> spec -> spec

Set the name of an estimator.

This setter function modifies an existing specification. The & operator is useful for chaining multiple modifications.

Example

Expand

-- Standard usage
spec' = name_ "baseline_regression" spec

-- Chaining with & operator
spec' = spec
    & name_ "baseline_regression"
    & vce_ RobustErrors

LHS lens

class HasLhs xsource (i :: k) spec | spec -> xsource, spec -> i where

Lens type class for left-hand side (dependent) variable.

Methods

lhsL :: Lens' spec (SpecVar xsource i)

lhs_ :: forall {k} xsource (i :: k) spec. HasLhs xsource i spec => SpecVar xsource i -> spec -> spec

Set the left-hand side (dependent) variable of a specification.

This setter function modifies an existing specification. The & operator is useful for chaining multiple modifications.

Example

Expand

-- Standard usage
spec' = lhs_ [lhsQ| car |] spec

-- Chaining with & operator
spec' = spec
    & lhs_ [lhsQ| car |]
    & vce_ RobustErrors

mapMaybeLhsVar :: forall {k} xsource (i :: k) spec. (Eq xsource, HasLhs xsource i spec) => (SpecVar xsource i -> Maybe (SpecVar xsource i)) -> spec -> Maybe spec

Transform the left-hand side variable with possible failure.

Applies a transformation that may fail (returning Nothing). If the transformation fails, the entire operation returns Nothing.

Useful for conditional transformations or validation.

RHS lens

class HasRhs xsource (i :: k) spec | spec -> xsource, spec -> i where

Lens type class for right-hand side (independent) variables.

Methods

rhsL :: Lens' spec (Rhs (SpecVar xsource i))

rhs_ :: forall {k} xsource (i :: k) spec. HasRhs xsource i spec => Rhs (SpecVar xsource i) -> spec -> spec

Set the right-hand side (independent) variables of a specification.

This setter function modifies an existing specification. The & operator is useful for chaining multiple modifications.

Example

Expand

-- Standard usage
spec' = rhs_ [rhsQ| dealValue + leverage |] spec

-- Chaining with & operator
spec' = spec
    & rhs_ [rhsQ| dealValue + leverage |]
    & vce_ RobustErrors

addToRhs_ :: forall {k} xsource (i :: k) spec. HasRhs xsource i spec => Rhs (SpecVar xsource i) -> spec -> spec

Add variables to the right-hand side of a specification.

Appends additional independent variables to the existing RHS using <>.

This function modifies an existing specification. The & operator is useful for chaining multiple modifications.

Example

Expand

-- Standard usage
spec' = addToRhs_ [rhsQ| sizeControl + industryControl |] spec

-- Chaining with & operator
spec' = spec
    & addToRhs_ [rhsQ| sizeControl |]
    & addToRhs_ [rhsQ| industryControl |]
    & vce_ RobustErrors

mapMaybeRhsVar :: forall {k} xsource (i :: k) spec. (Eq xsource, HasRhs xsource i spec) => (SpecVar xsource i -> Maybe (SpecVar xsource i)) -> spec -> Maybe spec

Transform right-hand side variables with possible failure.

Applies a transformation that may fail to each independent variable. If any transformation fails, the entire operation returns Nothing.

Conditions lens

class HasConditions xsource (i :: k) spec | spec -> xsource, spec -> i where

Lens type class for sample conditions.

Methods

conditionsL :: Lens' spec (Maybe (Conditions (SpecVar xsource i)))

conditions_ :: forall {k} xsource (i :: k) spec evar. (HasConditions xsource i spec, SpecVar xsource i ~ evar) => Maybe (Conditions evar) -> spec -> spec

Set the conditions for a specification.

Replaces any existing conditions with the provided ones.

This function modifies an existing specification. The & operator is useful for chaining multiple modifications.

Example

Expand

-- Standard usage
spec' = conditions_ [conditionsQ| publicTarget && completedDeal |] spec

-- Chaining with & operator
spec' = spec
    & conditions_ [conditionsQ| publicTarget |]
    & vce_ RobustErrors

andCondition_ :: forall {k} xsource (i :: k) spec evar. (HasConditions xsource i spec, SpecVar xsource i ~ evar) => Conditions evar -> spec -> spec

Add a condition to a specification with AND logic.

If there are existing conditions, combines them with ConditionAnd. Otherwise, sets the condition as the only restriction.

This function modifies an existing specification. The & operator is useful for chaining multiple modifications.

Example

Expand

-- Standard usage
spec' = andCondition_ (Condition publicTarget) spec

-- Chaining with & operator
spec' = spec
    & andCondition_ (Condition publicTarget)
    & andCondition_ (Condition completedDeal)
    & vce_ RobustErrors

replaceCondition_

Arguments

:: forall {k} xsource (i :: k) spec evar. (HasConditions xsource i spec, SpecVar xsource i ~ evar, Eq xsource)

=> (evar -> Bool)
A test to determine if the condition should be replaced

-> evar
The replacement condition

-> spec

-> spec

Replace specific conditions in a specification.

Traverses the condition tree and replaces any condition matching the test predicate with the target condition.

VCE lens

class HasVce xsource (i :: k) spec | spec -> xsource, spec -> i where

Lens type class for variance-covariance estimation (standard error calculation).

VCE specifies how standard errors should be computed: robust (heteroskedasticity-consistent), clustered (by one or more variables), or unadjusted (classical OLS).

Methods

vceL :: Lens' spec (Vce (SpecVar xsource i))

vce_ :: forall {k} xsource (i :: k) spec evar. (HasVce xsource i spec, SpecVar xsource i ~ evar) => Vce evar -> spec -> spec

Set the variance-covariance estimation method for a specification.

Specifies how standard errors should be computed in the regression.

This function modifies an existing specification. The & operator is useful for chaining multiple modifications.

Example

Expand

-- Standard usage with robust errors
spec' = vce_ RobustErrors spec

-- Use clustered standard errors
spec' = vce_ [clusterQ| firmId |] spec

-- Chaining with & operator
spec' = spec
    & lhs_ [lhsQ| car |]
    & rhs_ [rhsQ| dealValue + leverage |]
    & vce_ [clusterQ| firmId |]

Absorb lens

class HasAbsorb xsource (i :: k) spec | spec -> xsource, spec -> i where

Lens type class for absorbed fixed effects.

Methods

absorbL :: Lens' spec (Set (SpecVar xsource i))

absorb_ :: forall {k} xsource (i :: k) spec. HasAbsorb xsource i spec => Set (SpecVar xsource i) -> spec -> spec

Set the absorbed fixed effects for a specification.

Replaces any existing absorbed fixed effects with the provided ones.

This function modifies an existing specification. The & operator is useful for chaining multiple modifications.

Example

Expand

-- Standard usage
spec' = absorb_ [absorbQ| industry year |] spec

-- Chaining with & operator
spec' = spec
    & lhs_ [lhsQ| car |]
    & rhs_ [rhsQ| dealValue + leverage |]
    & absorb_ [absorbQ| industry year |]
    & vce_ [clusterQ| firmId |]

mapMaybeAbsorbVar :: forall {k} xsource (i :: k) spec. (Eq xsource, Ord xsource, HasAbsorb xsource i spec) => (SpecVar xsource i -> Maybe (SpecVar xsource i)) -> spec -> Maybe spec

Transform absorbed fixed effects with possible failure.

Applies a transformation that may fail to each absorbed variable. If any transformation fails, the entire operation returns Nothing.

Useful for filtering fixed effects or conditional transformations.